move load_json and cancel requests methods to base class
