+++ bento_etl/loaders/base.py
@@ -1,6 +1,14 @@
+from asyncio import Task
+import asyncio
+import json
 from logging import Logger
+from fastapi import status
+from httpx import AsyncClient
+import httpx
 
 from bento_etl.config import Config
+from bento_etl import authz
+
 
 __all__ = ["BaseLoader"]
 
@@ -19,6 +27,40 @@ class BaseLoader:
 
     def load(self, data):
         pass
+    
+    
+    async def _load_json(self, data: json):
+        load_requests = []
+        limits = httpx.Limits(max_keepalive_connections=20, max_connections=len(data))
+        headers = {
+            "Authorization": authz.get_bearer_token_from_config(self.config)
+        }
+
+        async with AsyncClient(
+            limits=limits, verify=self.config.bento_validate_ssl, headers=headers
+        ) as client:
+            try:
+                for index in range(0, len(data), self.batch_size):
+                    batch = data[index : index + self.batch_size]
+                    request = asyncio.ensure_future(self.send_request(client, batch))
+                    load_requests.append(request)
+                await asyncio.gather(*load_requests)
+            except Exception as ex:
+                self.logger.warning("Cancelling all uploads")
+                self.cancel_all_requests(load_requests)
+                raise ex
+
+    async def _send_json_data(self, client: AsyncClient, data: json):
+        response = await client.post(self.load_url, json=data)
+
+        if response.status_code != status.HTTP_204_NO_CONTENT:
+            error_message = f"Upload to Katsu failed with status code {response.status_code}"
+            self.logger.error(error_message)
+            raise Exception(error_message)
+
+    def cancel_all_requests(self, requests: list[Task]):
+        for request in requests:
+            request.cancel()
 
 
 # TODO: implement loaders for phenopackets and experiments
